requirements.txt
streamlit==0.61.0
Pillow==7.1.2
numpy==1.18.1
Keras==2.4.3